// frontend/lib/api/generated/model/pagePenalidadOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { PenalidadOut } from './penalidadOut';

export interface PagePenalidadOut {
  items: PenalidadOut[];
  total: number;
  page: number;
  page_size: number;
}